loader: Make loader into cell and provide to hive
This commit changes the loader into a cell. The Datapath object depends
on it and re-distributes it still (for now). All locations that created
a loader now reuse the loader instance from hive. The exception being
tests which did not yet use hive.
